requirements.txt
Flask
hanging_threads
matplotlib
numpy
pandas
pyserial
python-dateutil
python-dotenv
pytz
requests
scipy
skyfield
uwsgi
